Fix syntax issues.
Kramdown (the markdown processor we use) is very picky when it comes to markdown lists.
To render the list correctly, the content of each item has to be
indented by 3 spaces. As a general rule, you should indent the content
as much as the first character in the list item, e.g.

1. This is one item

   Content should be indented here