[Entity Schema] Render the schema on entity pages using the new spec-renderer and pulling the information from Konnect.
It works similarly to how the api specs work in the docs.
We keep a file app/_data/konnect_oas_data.json with the products'
information we have in Konnect so we can retrieve their API specs.

The file app/_data/konnect_product_ids.yml contains a mapping
of each API spec and its correpsonding product id in Konnect. This
provides a way to access and filter them easily.

The schema is rendered using a vue component that requires 3 arguments:
* konnect product id
* the version id of the specific product
* the path to the schema in the spec

The platform pulls this information from the different configuration
files based on the following metadata in a page's frontmatter:

```
schema:
  api: gateway/admin-ee
  path: /schemas/Consumer
```

The key `api` maps to the following entry in app/_data/konnect_product_ids.yml
```
gateway:
  admin-ee: 'konnect_product_id'
```

And the corresponding version information comes from
app/_data/konnect_oas_data.json