Add broken-link-checker
It runs in three different ways:
* Full site scan
* Run a check against the pages generated by specific source files
* PR mode: same as the above, but it pulls the list of files from the
  pull request.

It relies on a file generated by the site after build, which contains a
hash that maps all the source files used by Jekyll to generated pages
and the corresponding URLs.

The check can then fetch the urls it needs to check based on the source
files provided (either manually or pulled from the PR) and run
broken-link-checker against those.