Change the way versioning works for reference pages
Only reference pages that have `vesioned: true` in their frontmatters
will be versioned, the rest will uses the latest version.

If the min_version of a page is greater than the latest version, it will
not be generated in production, and will only be available in dev and
preview apps in `/url/unpublished/`.

Handling changes to an existing page for a new version requires the use
of {% if_version %} which we need to remove before the release.