Add `col_span` to landing pages rows
It allows to specify the number of columns a row can have, if not
present it falls back to the number of blocks.

Useful for setting generating layouts with X numbers of columns with Y
number of blocks, where Y is less than X, e.g. generating a 3 column
layout and having only one card.