Fix Accordions in how-tos
There was a bug where the last prereq was being expanded even though
default=0 was set.
This was caused by the way we were handling the topology-switch,
clicking every prereq and cleanup item on change.

A custom event solves this, the accordion listens to it, and updates
itself if the target of the event is its element.