Invalid OAS specs (#1684)
* TDX-5484 Invalid OAS specs

Invalid OAS specs are now permitted. Validation methodology is explained to assist users in resolving validation errors locally. Additionally, `validation_messages` are persisted to aid users in resolution. The validation status and messages are surfaced in the UX, but as most users are automating the API catalog, there is a DX-focus in docs.

* Apply suggestions from code review

---------

Co-authored-by: Diana <75819066+cloudjumpercat@users.noreply.github.com>