Added note about DNS records needing to be public. (#2108)
We had a support case where the Dev Portal names was no longer occurring because the SSL cert expired, and it expired because the DNS records were made private and not publicly seen which meant the ACME challenge for the SSL certificate was not able to pass the validation checks. Currently, the CNAME and CAA DNS records must be public so Konnect can resolve them.